Fixed edge case in file counts where sample is substring of another sample
The fix uses find with regex. The count_file_regex is on its own line as
we first need to resolve the bash variable in a way compatible with both
OS X and linux, as behaviour is a little different for resolving variables
(using extended -regextype or -E for linux and mac respectively).